IMPALA-5092 Add support for VARCHAR in Kudu tables
KUDU-1938 added VARCHAR column type support to Kudu. This commit adds
support for Kudu's VARCHAR type to Impala as well.

VARCHAR length is enforced in the Kudu client, Impala doesn't have to
enforce it further. The length of the VARCHAR is given in characters
(UTF8) so a VARCHAR(n) is stored in "n" to "4n" bytes.

The values are truncated in predicate pushdown also as the truncation is
handled in client-side.

TODO: add tests for the scenario where the primary key is varchar +
predicates and partitioning

Testing:

* performed manual testing
* added automated tests that passed

Change-Id: I0d4959410fdd882bfa980cb55e8a7837c7823da8